requirements.txt
fastapi
uvicorn
pydantic>=2
google-generativeai>=0.7.2
youtube-transcript-api
yt-dlp
faster-whisper
ffmpeg-python
python-multipart
requests
tqdm
openai

